Claim modifiers refactor
* Improved JavaDoc
* Moved to new package
