+++ oidc-idp/src/main/java/cz/muni/ics/oidc/server/PerunUserInfoRepository.java
@@ -7,8 +7,8 @@ import com.google.common.cache.CacheBuilder;
 import com.google.common.cache.CacheLoader;
 import com.google.common.cache.LoadingCache;
 import com.google.common.util.concurrent.UncheckedExecutionException;
-import cz.muni.ics.oidc.server.claims.ClaimValueModifier;
-import cz.muni.ics.oidc.server.claims.ClaimValueModifierInitContext;
+import cz.muni.ics.oidc.server.claims.modifiers.ClaimValueModifier;
+import cz.muni.ics.oidc.server.claims.modifiers.ClaimValueModifierInitContext;
 import cz.muni.ics.oidc.server.claims.PerunCustomClaimDefinition;
 import cz.muni.ics.oidc.server.connectors.PerunConnector;
 import cz.muni.ics.oidc.models.RichUser;
